Rename POI test for some weird pytest failures
